[수달] 2단계 재화 미션 제출합니다. (#39)
* docs: todo list 추가

* refactor: 상수 분리 및 패키지 변경

* refactor: xml 디자인 시안과 동일화

- 최근 본 목록 위의 제목 구현
- 상품 목록의 아이템이 대칭되게 구현
- 최근 본 상품과 상품 목록의 사이에 위치한 구분선 구현

* fix(recentProduct): 최근 상품이 서버와 유저에 관계 없이 노출되는 문제 해결

- DB에 serverUrl과 userEncoding값을 가지는 칼럼을 추가해 구분하도록 수정

* refactor(ui): 스켈레톤 UI가 꺼지는 시점 수정

- 2초 후 스켈레톤 UI가 꺼지도록 하는 것이 아닌, 상품이 로드됐을 때 꺼지도록 수정

* refactor: toolbar 구현

* fix: 장바구니 목록 아이템의 수량조절 + 버튼이 표시되지 않던 현상 수정

* refactor(repository): productRepository retrofit 적용

* refactor: 장바구니관련 통신 라이브러리 Retrofit으로 변경

* fix: dialog 에서 수량 조절 + 버튼이 나오지 않던 현상 수정

* fix: 메인 상품 목록에서 수량조절을 해서 0이 되었을 때 장바구니에서 제거되지 않는 현상 수정

* refactor(cart): cartPoduct를 관리하는 cart 일급컬렉션 추가

-

* fix(cart): 장바구니의 상품 금액 총합이 적용되지 않는 문제 해결 및 cart 관리 객체 적용

* docs: 기능 리스트 적용 사항 업데이트

* feat(cart): 장바구니 전체 선택 기능 추가

* fix(cart): 장바구니 제품 삭제시 앱 크래시 문제 해결

- cart 캐시를 담고 있는 객체를 관리하지 않는 문제 (삭제를 적용하지 않은 문제)
- view에서 삭제가 클릭되었을 때 삭제된 cartItem의 id가 아닌 cart의 productId를 반환하던 문제

* feat(order): 주문 관련 모델 추가

* feat(order): 주문 화면 MVP 임시구현

- 테스트 및 스켈레톤 코드 추가

* test(orderDetail): 주문 상세 화면 MVP 테스트 추가

* refactor: glide bindingAdapter 구현

* feat(cart): 선택된 장바구니 상품들로 이루어진 장바구니를 반환하는 기능구현

* feat(cart): Order 페이지로 이동하는 기능 구현

* feat(order): 주문 관련 기능 추가

- Repository 작성
- 전체 조회시 주문 정보의 일부만 담고 있는 OrderSummary 추가
- 그 외 Order, OrderProduct 추가

* feat(order): 주문페이지 구현

* refactor(BindingAdapter): 불필요한 중복코드 제거

* refactor(test): 테스트 abstract class 추가

* feat(order): 주문 내역 상세 화면 기능 추가

* refactor(order): 주문 전 상품을 담는 장바구니 이름 변경

* feat(products): 상품 목록을 일부만 가져오는 기능 작성

* refactor(package): 흩어져 있던 데이터 계층의 모델들을 데이터 계층으로 이동

* refactor(package): 데이터 계층의 모델 패키지 상세화

* refactor(cart): 장바구니 목록을 일부만 가져오는 기능 작성

* feat(order): 할인정책을 가지고와서 계산하여 표시하는 기능 구현

* refactor(product): 전체 기능 조회 삭제, 일부 조회 적용

- API 명세 변경으로 인한 전체 기능 조회가 사용불가해짐에 따라 변경
- 제품 일부 조회시 페이지네이션을 받아오는 부분도 적용 필요

* refactor(product): 페이지네이션 개선

- 서버로부터 받아온 pagination 적용

* fix(products): 장바구니에서 개수가 변경되었을 때 적용이 되지 않던 문제 해결

* fix(cart): 선택된 상품 총합 가격을 페이지내에서만 계산하는 현상 수정

* refactor: 불필요 코드 제거

* refactor(cart): 도메인 로직 분리

* refactor(doamin): 도메인 패키지 분리

* refactor(domain): 장바구니 도메인 로직 수정 및 테스트코드 추가

- 장바구니 제품의 개수를 설정하는 로직 수정
- 장바구니 isXXX이라는 네이밍을 가지는 메서드를 프로퍼티로 변경

* refactor: ui 로직에서 Thread를 관리하는 로직 제거

* refactor(ui): 스켈레톤 ui xml 코드를 분리

* refactor(retrofit): retrofit을 관리하는 로직 분리

* refactor(retrofit): retrofit callback을 관리하는 로직 분리

* refactor: presenter injector 작성

* refactor: repostory 위치 변경

---------

+++ app/src/main/java/woowacourse/shopping/model/CartProductState.kt
@@ -5,8 +5,8 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class CartProductState(
-    val id: Int,
-    val productId: Int,
+    val id: Long,
+    val productId: Long,
     val productImageUrl: String,
     val productName: String,
     val productPrice: Int,
